--- RangeBarSample/res/drawable/material_button.xml
@@ -1,57 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<selector xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <!-- pressed state -->
-    <item android:state_pressed="true">
-        <inset xmlns:android="http://schemas.android.com/apk/res/android"
-               android:insetLeft="@dimen/button_inset_horizontal_material"
-               android:insetTop="@dimen/button_inset_vertical_material"
-               android:insetRight="@dimen/button_inset_horizontal_material"
-               android:insetBottom="@dimen/button_inset_vertical_material">
-            <shape android:shape="rectangle">
-                <corners android:radius="@dimen/control_corner_material"/>
-                <solid android:color="@color/accent"/>
-                <padding android:left="@dimen/button_padding_horizontal_material"
-                         android:top="@dimen/button_padding_vertical_material"
-                         android:right="@dimen/button_padding_horizontal_material"
-                         android:bottom="@dimen/button_padding_vertical_material"/>
-            </shape>
-        </inset>
-    </item>
-
-    <!-- focused state -->
-    <item android:state_focused="true">
-        <inset xmlns:android="http://schemas.android.com/apk/res/android"
-               android:insetLeft="@dimen/button_inset_horizontal_material"
-               android:insetTop="@dimen/button_inset_vertical_material"
-               android:insetRight="@dimen/button_inset_horizontal_material"
-               android:insetBottom="@dimen/button_inset_vertical_material">
-            <shape android:shape="rectangle">
-                <corners android:radius="@dimen/control_corner_material"/>
-                <solid android:color="@color/accent"/>
-                <padding android:left="@dimen/button_padding_horizontal_material"
-                         android:top="@dimen/button_padding_vertical_material"
-                         android:right="@dimen/button_padding_horizontal_material"
-                         android:bottom="@dimen/button_padding_vertical_material"/>
-            </shape>
-        </inset>
-    </item>
-
-    <!-- normal state -->
-    <item>
-        <inset xmlns:android="http://schemas.android.com/apk/res/android"
-               android:insetLeft="@dimen/button_inset_horizontal_material"
-               android:insetTop="@dimen/button_inset_vertical_material"
-               android:insetRight="@dimen/button_inset_horizontal_material"
-               android:insetBottom="@dimen/button_inset_vertical_material">
-            <shape android:shape="rectangle">
-                <corners android:radius="@dimen/control_corner_material"/>
-                <!--<solid android:color="@color/NORMAL_STATE_COLOR" />-->
-                <padding android:left="@dimen/button_padding_horizontal_material"
-                         android:top="@dimen/button_padding_vertical_material"
-                         android:right="@dimen/button_padding_horizontal_material"
-                         android:bottom="@dimen/button_padding_vertical_material"/>
-            </shape>
-        </inset>
-    </item>
-</selector>
