+++ RangeBarSample/res/layout/color_picker_dialog.xml
@@ -34,7 +34,7 @@
             android:layout_gravity="center"
             android:gravity="center" />
 
-        <com.appyvet.rangebarsample.colorpicker.ColorPickerPalette
+        <com.github.michiel.rangebarsample.colorpicker.ColorPickerPalette
             android:id="@+id/color_picker"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
