+++ RangeBarSample/src/com/github/michiel/rangebarsample/Component.java
@@ -1,5 +1,5 @@
 
-package com.appyvet.rangebarsample;
+package com.github.michiel.rangebarsample;
 
 /**
  * Enumeration of the components to be changed in colorChanged. Only needed for
