+++ RangeBarSample/src/com/github/michiel/rangebarsample/colorpicker/Utils.java
@@ -13,13 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *******************************************************************************/
-package com.appyvet.rangebarsample.colorpicker;
-
-import com.appyvet.rangebarsample.R;
+package com.github.michiel.rangebarsample.colorpicker;
 
 import android.content.Context;
 import android.content.res.Configuration;
 import android.graphics.Color;
+
+import com.github.michiel.rangebarsample.R;
 
 /**
  *
