+++ rangebar/build.gradle
@@ -1,21 +1,11 @@
 apply plugin: 'com.android.library'
-apply plugin: 'maven'
-apply plugin: 'signing'
-apply from: '../maven_push.gradle'
-
-version = "1.3"
-group = "com.appyvet"
-dependencies {
-    compile 'com.android.support:support-v4:23.1.0'
-}
 android {
-
-    compileSdkVersion 23
-    buildToolsVersion "23.0.1"
+    compileSdkVersion 24
+    buildToolsVersion "23.0.3"
 
     defaultConfig {
         minSdkVersion 12
-        targetSdkVersion 23
+        targetSdkVersion 24
     }
 
     sourceSets {
@@ -30,24 +20,6 @@ android {
 repositories {
     mavenCentral()
 }
-
-//signing {
-//    required { has("release") && gradle.taskGraph.hasTask("uploadArchives") }
-//    sign configurations.archives
-//}
-
-task publishLocal(type: Upload) {
-    configuration = configurations.archives
-    repositories.mavenDeployer {
-        repository(url: "file://$buildDir/repo")
-    }
-
-    task sourcesJar(type: Jar) {
-        classifier = 'sources'
-        from android.sourceSets.main.java.srcDirs
-    }
-
-    task generateJavadocs(type: Javadoc) {
-        source = android.sourceSets.main.java.srcDirs
-    }
+dependencies {
+    compile 'com.android.support:support-v4:24.0.0'
 }
