--- rangebar/src/com/appyvet/rangebar/IRangeBarFormatter.java
@@ -1,10 +1,0 @@
-package com.appyvet.rangebar;
-
-/**
- * Created by franciscomorais on 24/06/15.
- */
-public interface IRangeBarFormatter {
-
-    String format(String value);
-
-}
