+++ rangebar/src/com/github/michiel/rangebar/PinView.java
@@ -1,17 +1,4 @@
-/*
- * Copyright 2014, Appyvet, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this work except in compliance with the License.
- * You may obtain a copy of the License in the LICENSE file, or at:
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" 
- * BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language 
- * governing permissions and limitations under the License. 
- */
-
-package com.appyvet.rangebar;
+package com.github.michiel.rangebar;
 
 import android.content.Context;
 import android.content.res.Resources;
@@ -24,6 +11,9 @@ import android.graphics.drawable.Drawable;
 import android.support.v4.content.ContextCompat;
 import android.util.TypedValue;
 import android.view.View;
+
+import com.github.michiel.rangebar.util.IRangeBarFormatter;
+import com.mick.rangebar.R;
 
 /**
  * Represents a thumb in the RangeBar slider. This is the handle for the slider
