+++ rangebar/src/com/github/michiel/rangebar/util/IRangeBarFormatter.java
@@ -1,0 +1,5 @@
+package com.github.michiel.rangebar.util;
+
+public interface IRangeBarFormatter {
+	String format(String value);
+}
